refactor(Topic): seperate controller & service
